arithmetic-wild: SAE numerical fixes + analysis scripts
- sae_loader: load SAEs in float32 (bf16 interchange overflowed to NaN)
- llama.py _sae_intervene: cast acts to SAE dtype; norm-clamp guard (cap 8x clean
  cf-norm, fall back to cf if non-finite) to stop reconstruction error compounding
  across 32 layers
- gen_addition: set pad_token (tokenizer had none)
- attribute_arith: per-task non-finite-loss detection
- analyze_arith / dbg_das_grad: layer-importance profiles + DAS gradient diagnostic

+++ src/learning_to_attribute/sae_loader.py
@@ -28,5 +28,7 @@ class LlamaScopeSAE:
         return (g.to(self.W_dec.dtype) @ self.W_dec.T) / self.s
 
 
-def load_llama_scope_saes(repo, n_layers, device, dtype=torch.bfloat16):
+def load_llama_scope_saes(repo, n_layers, device, dtype=torch.float32):
+    # float32 by default: the bf16 interchange overflowed to NaN when feature
+    # deltas are injected across all 32 layers during training.
     return {li: LlamaScopeSAE(repo, li, device, dtype) for li in range(n_layers)}
